streamlit/requirements.txt
streamlit
pandas
matplotlib
plotly.express